requirements.txt
keras-nlp
keras==2.4.3
numpy==1.19.5
scikit-learn==0.24.2
transformers
tensorflow-gpu==2.4.1
torch